refactor: isolate shared configuration errors
