+++ quadletgen/storage_model.py
@@ -11,7 +11,7 @@ from dataclasses import dataclass, field
 from enum import Enum
 from typing import Literal, NoReturn, TypeAlias
 
-from .model import ConfigError
+from .errors import ConfigError
 
 
 NAME_RE = re.compile(r"^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$")
