+++ quadletgen/storage_parser.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 from collections.abc import Mapping
 from typing import NoReturn
 
-from .model import ConfigError
+from .errors import ConfigError
 from .storage_model import (
     DirectoryStorage,
     ExistingZfsStorage,
